python/requirements.txt
requests>=2.0.0
beautifulsoup4>=4.0.0
ingredient_parser_nlp